refactor: 폴더 정리 및, Modal에서 Backdrop 분리, Contents 삭제
- 컴포넌트들을 components/components 폴더 내에 위치 시킴
- 각각의 Modal타입 컴포넌트에서 타입에 맞게 Backdrop과 modal-contents 를 관리하도록 함
