+++ components/src/App.tsx
@@ -1,17 +1,8 @@
 import { useRef, useState } from 'react';
 import './App.css';
-import styled from 'styled-components';
 
 import { BottomModal, Modal } from './lib';
 import { usePosition } from './lib/hooks/';
-
-const ContentsInner = styled.div<{ backgroundColor?: string; width?: string; height?: string }>`
-  padding: 2rem 1.5rem;
-  box-sizing: border-box;
-  background-color: ${({ backgroundColor }) => backgroundColor || '#fff'};
-  border-radius: 0.625rem;
-  width: ${({ width }) => width || 'fit-contents'};
-`;
 
 function App() {
   const [openCenterModal, setOpenCenterModal] = useState(false);
@@ -26,16 +17,12 @@ function App() {
       <button onClick={() => setOpenBottomModal(true)}> bottom modal open</button>
       <button onClick={() => setOpenToastModal(true)}> toast modal open</button>
       <Modal type="bottom" openModal={openBottomModal} setOpenModal={setOpenBottomModal}>
-        <ContentsInner>
-          <h1>Bottom Modal</h1>
-          <BottomModal.button isCloseModal={true}>close</BottomModal.button>
-        </ContentsInner>
+        <h1>Bottom Modal</h1>
+        <BottomModal.button isCloseModal={true}>close</BottomModal.button>
       </Modal>
       <Modal type="center" openModal={openCenterModal} setOpenModal={setOpenCenterModal}>
-        <ContentsInner>
-          <h1>Center Modal</h1>
-          <Modal.button isCloseModal={true}>close button</Modal.button>
-        </ContentsInner>
+        <h1>Center Modal</h1>
+        <Modal.button isCloseModal={true}>close button</Modal.button>
       </Modal>
       <Modal
         type="toast"
@@ -43,10 +30,12 @@ function App() {
         setOpenModal={setOpenToastModal}
         position={position}
         isNeedAnimation={true}
+        backgroundColor={{ modal: 'rgb(248, 255, 188)' }}
+        contentsPadding="1rem 0.875rem"
       >
-        <ContentsInner backgroundColor=" rgb(248, 255, 188)" width="300px">
-          <h3>toast modal</h3>
-        </ContentsInner>
+        <div style={{ width: '300px', height: '2rem', textAlign: 'center' }}>
+          <h2>toast modal</h2>
+        </div>
       </Modal>
       <div ref={positionRef} id="toast-modal-position">
         toast modal position
