+++ components/src/lib/components/CloseButtonIcon.tsx
