+++ components/src/lib/components/index.tsx
@@ -1,0 +1,2 @@
+export { default as Modal } from './Modal';
+export { default as BottomModal } from './BottomModal';
