+++ components/src/lib/constants/modal.ts
@@ -1,3 +1,5 @@
 export const BASIC_BOTTOM_MODAL_ANIMATION_DURATION = 1000;
 export const NOW = 0;
 export const BASIC_TOAST_MODAL_ANIMATION_DURATION = 3000;
+export const BASIC_TOAST_DURATION = 6000;
+export const CONTENTS_CLASS_NAME = 'modal-contents';
