+++ components/src/lib/hooks/index.ts
@@ -1,4 +1,5 @@
+export { default as useBottomModalAnimation } from './useBottomModalAnimation';
 export { default as useBottomModalContext } from './useBottomModalContext';
 export { default as useModalContext } from './useModalContext';
 export { default as usePosition } from './usePosition';
-export { default as useBottomModalAnimation } from './useBottomModalAnimation';
+export { default as useToastModalAnimation } from './useToastModalAnimation';
