+++ components/src/lib/index.ts
@@ -1,7 +1,4 @@
 export * from './contexts';
 export * from './hooks';
 export * from './types/modal';
-export { default as Modal } from './Modal';
-export { default as BottomModal } from './Modal/BottomModal';
-export { default as CenterModal } from './Modal/CenterModal';
-export { default as TostModal } from './Modal/ToastModal';
+export * from './components';
